Initial re-implementation of the 'tracker' program
Track internet use time by user, allowing a sysadmin to specify maximum
time limits (per user) in /var/log/tracker/%s (where %s is the login name).

The file needs to be written with just one number: the number of seconds
per day that the user is allowed for a connection.  The tracker will
update the file to contain how many seconds are left, and when the last
time the user was logged on was, along with full name and time left
information, so it turns the file into something like this:

	[torvalds@nehalem ~]$ cat /var/log/tracker/torvalds
	2600 441 1222822497
	Linus Torvalds
	0:35:59 left

which means that user 'torvalds' is allowed 2600 seconds of internet per
day (that's the only part that the admin needs to fill in, and can be
updated by just writing a new number to the file), has used 441 seconds
today, and was last seen logged in at 1222822497 (seconds since epoch).

The full name and time left is purely for visualizers, ie some trivial
tcl/tk app that reads this file could/should just ignore the real data,
and pick up that "time left" part.  You could just parse the raw data,
of course, but since I'm lousy at tcl/tk, I think it's easier to make
the C parts generate all of it ;)

The "per day" is approximate - tracker will clear the 'used time'
whenever more than eight hours have passed since last login.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>